small fix for adc read delay

+++ src/incur.h
@@ -32,7 +32,7 @@ class incur : public ofxMidiListener {
     TerminalListener consoleListener;
     void onCharacterReceived(KeyListenerEventData& e);
     // analog inputs
-    int adcDelay; 
+    float adcDelay; 
     MCP a2d;
     MCP_DESC chip = MCP3008;
     vector<int> analogIn;
